Refactor for structure for ValidationErrors.
