--- source/Energinet.DataHub.MeteringPoints.Application/Validation/Rules/MeteringPointTypeRequiredValidationError.cs
@@ -1,14 +1,0 @@
-﻿using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
-
-namespace Energinet.DataHub.MeteringPoints.Application.Validation.Rules
-{
-    public class MeteringPointTypeRequiredValidationError : ValidationError
-    {
-        public MeteringPointTypeRequiredValidationError(string meteringPointType)
-        {
-            MeteringPointType = meteringPointType;
-        }
-
-        public string MeteringPointType { get; }
-    }
-}
