+++ source/Energinet.DataHub.MeteringPoints.Application/Validation/Rules/MeteringPointTypeValidRule.cs
@@ -14,6 +14,7 @@
 
 using System.Data;
 using System.Linq;
+using Energinet.DataHub.MeteringPoints.Application.Validation.ValidationErrors;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 using FluentValidation;
