+++ source/Energinet.DataHub.MeteringPoints.Application/Validation/ValidationErrors/OccurenceRequiredValidationError.cs
@@ -14,15 +14,9 @@
 
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Application.Validation.Rules
+namespace Energinet.DataHub.MeteringPoints.Application.Validation.ValidationErrors
 {
-    public class GsrnNumberMustBeValidValidationError : ValidationError
+    public class OccurenceRequiredValidationError : ValidationError
     {
-        public GsrnNumberMustBeValidValidationError(string gsrnNumber)
-        {
-            GsrnNumber = gsrnNumber;
-        }
-
-        public string GsrnNumber { get; }
     }
 }
